Implementa tratamento de erros, exclusões, delete cascade dos livros;

+++ app/Http/Controllers/AssuntoController.php
@@ -4,7 +4,11 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\AssuntoRequest;
 use App\Models\Assunto;
+use App\Models\Livro;
+use Illuminate\Database\QueryException;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Log;
 
 class AssuntoController extends Controller
 {
@@ -21,8 +25,34 @@ class AssuntoController extends Controller
 
     public function store(AssuntoRequest $request)
     {
-        Assunto::create($request->all());
-        return redirect()->route('assuntos.index')->with('success', 'Assunto adicionado com sucesso!');
+        try {
+            DB::beginTransaction();
+
+            Assunto::create($request->all());
+
+            DB::commit();
+
+            return redirect()->route('assuntos.index')->with('success', 'Assunto adicionado com sucesso!');
+        } catch (QueryException $e) {
+            DB::rollBack();
+            Log::info($e->getMessage());
+
+            return redirect()
+                ->back()
+                ->withErrors(['error' => 'Erro ao salvar o assunto no banco de dados.']);
+        } catch (\PDOException $e) {
+            DB::rollBack();
+            Log::info($e->getMessage());
+
+            return redirect()
+                ->back()
+                ->withErrors(['error' => 'Erro de conexão com o banco de dados.']);
+        } catch (\Exception $e) {
+            DB::rollBack();
+            Log::info($e->getMessage());
+
+            return redirect()->back()->withErrors(['error' => 'Ocorreu um erro ao tentar adicionar o assunto.']);
+        }
     }
 
     public function edit(Assunto $assunto)
@@ -32,8 +62,34 @@ class AssuntoController extends Controller
 
     public function update(AssuntoRequest $request, Assunto $assunto)
     {
-        $assunto->update($request->all());
-        return redirect()->route('assuntos.index')->with('success', 'Assunto atualizado com sucesso!');
+        try {
+            DB::beginTransaction();
+
+            $assunto->update($request->all());
+
+            DB::commit();
+
+            return redirect()->route('assuntos.index')->with('success', 'Assunto atualizado com sucesso!');
+        } catch (QueryException $e) {
+            DB::rollBack();
+            Log::info($e->getMessage());
+
+            return redirect()
+                ->back()
+                ->withErrors(['error' => 'Erro ao atualizar o assunto no banco de dados.']);
+        } catch (\PDOException $e) {
+            DB::rollBack();
+            Log::info($e->getMessage());
+
+            return redirect()
+                ->back()
+                ->withErrors(['error' => 'Erro de conexão com o banco de dados.']);
+        } catch (\Exception $e) {
+            DB::rollBack();
+            Log::info($e->getMessage());
+
+            return redirect()->back()->withErrors(['error' => 'Ocorreu um erro ao tentar atualizar o assunto.']);
+        }
     }
 
     public function show(Assunto $assunto)
@@ -45,7 +101,12 @@ class AssuntoController extends Controller
 
     public function destroy(Assunto $assunto)
     {
+        // Se existir livro deste assunto
+        if ($assunto->livros()->exists()) {
+            return redirect()->route('assuntos.index')
+                ->withErrors(['error' => 'O assunto não pode ser deletado, pois está associado a um ou mais livros.']);
+        }
         $assunto->delete();
-        return redirect()->route('assuntos.index');
+        return redirect()->route('assuntos.index')->with('success', 'Assunto excluído com sucesso!');
     }
 }
